Fix(rust-tokenizer): increase integer width when converting hex literals (#4573)

+++ sqlglotrs/src/tokenizer.rs
@@ -458,7 +458,7 @@ impl<'a> TokenizerState<'a> {
         let text = self.extract_string(&end, false, token_type == self.token_types.raw_string, true)?;
 
         if let Some(b) = base {
-            if u64::from_str_radix(&text, b).is_err() {
+            if u128::from_str_radix(&text, b).is_err() {
                 return self.error_result(format!(
                     "Numeric string contains invalid characters from {}:{}",
                     self.line, self.start
